[FLINK-36066][runtime] Introduce the AdaptiveGraphManager component
