+++ flink-runtime/src/main/java/org/apache/flink/streaming/api/graph/StreamGraphHasherV2.java
@@ -148,6 +148,18 @@ public class StreamGraphHasherV2 implements StreamGraphHasher {
         return hasher.hash().asBytes();
     }
 
+    @Override
+    public boolean generateHashesByStreamNodeId(
+            int streamNodeId, StreamGraph streamGraph, Map<Integer, byte[]> hashes) {
+        StreamNode streamNode = streamGraph.getStreamNode(streamNodeId);
+        return generateNodeHash(
+                streamNode,
+                Hashing.murmur3_128(0),
+                hashes,
+                streamGraph.isChainingEnabled(),
+                streamGraph);
+    }
+
     /**
      * Generates a hash for the node and returns whether the operation was successful.
      *
